pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-fleet"
version = "0.5.0"
description = "A powerful multi-agent system for adaptive AI reasoning and automation with a clean, modular architecture"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
keywords = ["ai", "agents", "automation", "multi-agent", "adaptive", "AI", "reasoning", "autogen-sample", "autogen", "magentic-one"]
authors = [
  { name = "Zachary BENSALEM", email = "contact@qredence.ai" },
]

classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    # Core Framework
    "fastapi>=0.115.0",
    "pydantic>=2.5.2",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.24.0",
    "starlette>=0.41.3",
    "typer>=0.15.1",
    "click>=8.1.8",
    "packaging>=23.2",
    "typing-extensions>=4.8.0",
    # Networking & Async
    "aiofiles>=23.1.0",
    "aiohttp>=3.11.11,<4.0.0",
    "asyncio>=3.4.3",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "requests>=2.32.3",
    "websockets>=14.2",
    "python-socketio>=5.12.1",
    "sse-starlette>=2.2.1",
    "asyncer>=0.0.7",
    "nest-asyncio>=1.6.0",
    # Database & Storage
    "sqlalchemy[asyncio]>=2.0.28",
    "psycopg2-binary>=2.9.9",
    "redis>=5.2.1",
    "supabase>=2.3.1",
    # AI/ML Infrastructure
    "autogen-agentchat>=0.5.7,<0.6.0",
    "autogen-core>=0.5.7,<0.6.0",
    "autogen-ext[langchain,docker]>=0.5.7,<0.6.0",
    "autogen-ext>=0.5.7,<0.6.0", # Temporarily removed extras pulling openai-whisper
    "autogen-ext[openai,azure,gemini]>=0.5.7,<0.6.0",
    "autogen-ext[mcp,diskcache]>=0.5.7,<0.6.0",
    "autogen-ext[grpc]>=0.5.7,<0.6.0",
    "autogenstudio",
    "semantic-kernel>=1.19.0",
    "tiktoken>=0.8.0",
    # Model Providers
    "magentic-one-cli",
    "mcp>=1.6.0",
    "literalai>=0.1.201",
    # Azure Services
    "azure-ai-inference>=1.0.0b6", # For Azure AI Foundry
    "azure-ai-ml>=1.12.1", # For Azure AI Foundry
    "azure-ai-projects>=1.0.0b4", # For Azure AI Foundry
    "azure-core>=1.32.0",
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.19.0",
    "azure-keyvault-secrets>=4.9.0",
    "azure-search-documents==11.6.0b4",
    "azure-storage-blob>=12.24.0",
    # Data Processing
    "jsonlines>=4.0.0",
    "numpy>=1.26",
    "pandas>=2.2.3",
    "pytz>=2024.2",
    "scikit-learn>=1.6.1",
    "dataclasses-json>=0.6.7",
    # Visualization
    "matplotlib>=3.10.0",
    "pillow>=11.0.0",
    "seaborn>=0.13.0",
    # Web & Utilities
    "beautifulsoup4>=4.12.0",
    "bing>=0.31",
    "chainlit>=2.5.5",
    "docker>=7.1.0",
    "markitdown",
    "playwright>=1.20.0",
    "requests-html>=0.10.0",
    "tenacity>=9.0.0",
    "html2text>=2024.2.26",
    "pyjwt>=2.10.1",
    "pyyaml==5.3.1", # Using a specific version known to work with Python 3.12
    "psutil>=5.9.8",
    # Monitoring & Instrumentation
    "opentelemetry-instrumentation-fastapi>=0.44b0",
    "uptrace>=1.27.0",
    # Development Tools
    "rich>=13.9.4",
    "aioconsole>=0.8.1",
    "authlib>=1.4.0",
    "debugpy>=1.8.12",
    "ipykernel>=6.29.5",
    "puremagic>=1.28",
    "syncer>=2.0.3",
    "mem0ai>=0.1.86",
    # gRPC Support
    "grpcio>=1.70.0",
    "grpcio-tools>=1.70.0", # Matching grpcio version
    "absl-py>=2.1.0",
    # macOS specific (optional but included for compatibility)
    "macholib>=1.16.3; sys_platform == 'darwin'",
    "pytest>=8.3.4",
    "ruff>=0.11.10",
    "black>=25.1.0",
    "isort>=6.0.1",
]

[project.optional-dependencies]
test = [
    # Testing Framework
    "pytest>=8.0.2,<9.0.0",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.11.1",
    "pytest-sugar>=0.9.7",
    "pytest-xdist>=3.5.0",
    "coverage>=7.4.4",

    # HTTP Testing
    "httpx>=0.25.0",
    "respx>=0.20.2",

    # Code Quality
    "mypy>=1.5.0",
    "codecov>=2.1.13"
]

dev = [
    # Linting and Code Quality
    "ruff>=0.5.3",
    "black>=24.10.0",
    "isort>=5.13.2",
    "mypy>=1.5.0",

    # Testing
    "pytest>=8.0.2,<9.0.0",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.11.1",
    "pytest-sugar>=0.9.7",
    "pytest-xdist>=3.5.0",
    "coverage>=7.4.4",

    # Documentation
    "mkdocs>=1.5.3",
    "mkdocstrings[python]>=0.24.0",

    # Debugging
    "ipdb>=0.13.13",
    "rich>=13.9.4",
    "debugpy>=1.8.12"
]

docs = [
    "mkdocs>=1.5.3",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-material>=9.5.0",
    "pymdown-extensions>=10.7.0"
]

models = [
    # Local Models
    "ollama>=0.1.5",

    # Cloud Providers
    "autogen-ext[semantic-kernel-anthropic]>=0.5.7,<0.6.0",
    "autogen-ext[semantic-kernel-google]>=0.5.7,<0.6.0",
    "autogen-ext[semantic-kernel-ollama]>=0.5.7,<0.6.0",
    "autogen-ext[semantic-kernel-aws]>=0.5.7,<0.6.0",
    "autogen-ext[semantic-kernel-mistralai]>=0.5.7,<0.6.0",
    "autogen-ext[semantic-kernel-hugging-face]>=0.5.7,<0.6.0"
]

deploy = [
    "gunicorn>=23.0.0",
    "uvicorn[standard]>=0.24.0",
    "supervisor>=4.2.5",
    # Using docker SDK
    "docker>=7.0.0"
    # Note: docker-compose should be installed via pip install docker-compose
    # or via the Docker Desktop application
]

[tool.pytest.ini_options]
minversion = "8.0.0"
addopts = "-ra -q --cov=agentic_fleet --cov-report=term-missing"
testpaths = ["tests"]
python_files = "test_*.py"

[tool.coverage.run]
source = ["agentic_fleet"]
omit = [
    "tests/*",
    "*/__init__.py"
]

[tool.ruff]
line-length = 120
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP", "ASYNC"]
ignore = ["E501"]

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 120

[tool.hatch.build]
packages = ["src/agentic_fleet"]
exclude = [
    # Python
    "*.pyc",
    "__pycache__",
    "*.pyo",
    "*.pyd",
    "*.so",
    "*.dylib",
    "*.egg",
    "*.egg-info",
    # Build
    "build",
    "dist",
    "eggs",
    "sdist",
    "develop-eggs",
    # Directories
    "tests",
    "docs",
    "examples",
    ".git",
    ".github",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    # Config files
    ".gitignore",
    ".env*",
    "*.log",
    "tox.ini",
    ".coverage",
    "coverage.xml",
    # System
    ".DS_Store",
    "Thumbs.db",
]

[tool.hatch.version]
path = "src/agentic_fleet/__init__.py"

[tool.setuptools.packages.find]
exclude = [
    "tests*",
    "docs*",
    ".github*"
]

[tool.setuptools]
# Prevent accidental editable installations

[tool.setuptools.dynamic]
version = {attr = "agentic_fleet.__version__"}

# Warning for editable installations
[tool.setuptools.package-data]
"agentic_fleet" = [
    "EDITABLE_INSTALL_WARNING.txt"
]

[project.urls]
Homepage = "https://github.com/qredence/agentic-fleet"
Documentation = "https://github.com/qredence/agentic-fleet#readme"
Issues = "https://github.com/qredence/agentic-fleet/issues"
Source = "https://github.com/qredence/agentic-fleet"

[project.scripts]
agentic-fleet = "agentic_fleet.main:main"
agentic-fleet-chainlit = "agentic_fleet.chainlit_app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_fleet"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
